limiting size to 200

+++ philo/src/utils/pl_utl_atoi.c
@@ -6,7 +6,7 @@
 /*   By: erijania <erijania@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/21 15:40:06 by erijania          #+#    #+#             */
-/*   Updated: 2024/09/17 23:50:24 by erijania         ###   ########.fr       */
+/*   Updated: 2024/09/21 11:22:44 by erijania         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ static int	char_to_digit(char c)
 	return (-1);
 }
 
-int	pl_utl_atoi(const char *str)
+long	pl_utl_atoi(const char *str)
 {
-	int	ret;
-	int	sign;
+	long	ret;
+	int		sign;
 
 	sign = 1;
 	ret = 0;
